requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.11.0
apache-airflow-providers-common-compat==1.11.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.13.0
apache-airflow-providers-google==19.3.0
apache-airflow-providers-http==5.3.0
apache-airflow-providers-imap==3.9.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-sqlite==4.1.0
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.4
cattrs==25.3.0
certifi==2025.4.26
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
clickclick==20.10.2
colorama==0.4.6
colorful==0.5.8
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==46.0.3
db-dtypes==1.5.0
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email_validator==2.2.0
fastuuid==0.14.0
filelock==3.20.2
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.6.0
fsspec==2025.12.0
gcloud-aio-auth==5.4.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.6.1
gcsfs==2025.12.0
google-ads==28.4.1
google-analytics-admin==0.26.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.132.0
google-cloud-alloydb==0.6.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-automl==2.17.0
google-cloud-batch==0.19.0
google-cloud-bigquery==3.39.0
google-cloud-bigquery-datatransfer==3.20.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-build==3.34.0
google-cloud-compute==1.40.0
google-cloud-container==2.61.0
google-cloud-core==2.5.0
google-cloud-datacatalog==3.28.0
google-cloud-dataflow-client==0.10.0
google-cloud-dataform==0.7.0
google-cloud-dataplex==2.15.0
google-cloud-dataproc==5.23.0
google-cloud-dataproc-metastore==1.20.0
google-cloud-dlp==3.33.0
google-cloud-kms==3.7.0
google-cloud-language==2.18.0
google-cloud-logging==3.13.0
google-cloud-managedkafka==0.2.0
google-cloud-memcache==1.13.0
google-cloud-monitoring==2.28.0
google-cloud-orchestration-airflow==1.18.0
google-cloud-os-login==2.18.0
google-cloud-pubsub==2.34.0
google-cloud-redis==2.19.0
google-cloud-resource-manager==1.15.0
google-cloud-run==0.13.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.61.0
google-cloud-speech==2.35.0
google-cloud-storage==3.7.0
google-cloud-storage-control==1.8.0
google-cloud-storage-transfer==1.18.0
google-cloud-tasks==2.20.0
google-cloud-texttospeech==2.33.0
google-cloud-translate==3.23.0
google-cloud-videointelligence==2.17.0
google-cloud-vision==3.11.0
google-cloud-workflows==1.19.0
google-crc32c==1.8.0
google-genai==1.56.0
google-re2==1.1.20240702
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-gcp==0.2.2
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.10
immutabledict==4.2.2
importlib-metadata==6.11.0
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
limits==5.2.0
linkify-it-py==2.0.3
litellm==1.80.5
lockfile==0.12.2
looker_sdk==25.20.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.2.0
more-itertools==10.7.0
msgpack==1.1.2
multidict==6.4.4
numpy==2.4.0
oauthlib==3.3.1
openai==2.14.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-exporter-prometheus==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pandas-gbq==0.32.0
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.1
pluggy==1.6.0
prison==0.2.1
prometheus_client==0.23.1
propcache==0.3.1
proto-plus==1.27.0
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.11
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
pytest==9.0.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
ray==2.53.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.25.0
rsa==4.9.1
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.16.3
setproctitle==1.3.6
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.16.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.17.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.21.0
types-protobuf==6.32.1.20251210
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.4.0
virtualenv==20.35.4
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.0
zipp==3.21.0
